+++ ApiBundle/Tests/Transformer/StatusTransformerTest.php
@@ -3,6 +3,7 @@
 namespace OpenOrchestra\ApiBundle\Tests\Transformer;
 
 use Doctrine\Common\Collections\ArrayCollection;
+use OpenOrchestra\Backoffice\Security\ContributionActionInterface;
 use OpenOrchestra\BackofficeBundle\StrategyManager;
 use OpenOrchestra\BaseBundle\Tests\AbstractTest\AbstractBaseTestCase;
 use Phake;
@@ -14,20 +15,19 @@ use OpenOrchestra\ApiBundle\Transformer\StatusTransformer;
 class StatusTransformerTest extends AbstractBaseTestCase
 {
     protected $facadeClass = 'OpenOrchestra\ApiBundle\Facade\StatusFacade';
-    protected $authorizeStatusChangeManager;
     protected $authorizationChecker;
     protected $transformerManager;
     protected $groupContext;
     protected $transformer;
     protected $translator;
     protected $status;
+    protected $usageFinder;
 
     /**
      * Set up the test
      */
     public function setUp()
     {
-        $this->authorizeStatusChangeManager = Phake::mock('OpenOrchestra\BackofficeBundle\StrategyManager\AuthorizeStatusChangeManager');
         $this->authorizationChecker = Phake::mock('Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface');
         $this->groupContext = Phake::mock('OpenOrchestra\BaseApi\Context\GroupContext');
         $multiLanguagesChoiceManager = Phake::mock('OpenOrchestra\ModelInterface\Manager\MultiLanguagesChoiceManagerInterface');
@@ -41,7 +41,6 @@ class StatusTransformerTest extends AbstractBaseTestCase
 
         $statusId = 'StatusId';
 
-        Phake::when($this->authorizationChecker)->isGranted(Phake::anyParameters())->thenReturn(true);
         Phake::when($this->status)->getId()->thenReturn($statusId);
         Phake::when($router)->generateRoute(Phake::anyParameters())->thenReturn('route');
         Phake::when($statusRepository)->find(Phake::anyParameters())->thenReturn($this->status);
@@ -53,10 +52,10 @@ class StatusTransformerTest extends AbstractBaseTestCase
         Phake::when($groupContext)->hasGroup(Phake::anyParameters())->thenReturn(true);
 
         $this->usageFinder = Phake::mock('OpenOrchestra\Backoffice\UsageFinder\StatusUsageFinder');
+        Phake::when($this->usageFinder)->hasUsage(Phake::anyParameters())->thenReturn(false);
 
         $this->transformer = new StatusTransformer(
             $this->facadeClass,
-            $this->authorizeStatusChangeManager,
             $roleRepository,
             $multiLanguagesChoiceManager,
             $this->translator,
@@ -87,7 +86,9 @@ class StatusTransformerTest extends AbstractBaseTestCase
 
         $roles->add($role);
 
-        Phake::when($this->authorizeStatusChangeManager)->isGranted(Phake::anyParameters())->thenReturn($isGranted);
+        Phake::when($this->authorizationChecker)->isGranted($this->status, $document)->thenReturn($isGranted);
+        Phake::when($this->authorizationChecker)->isGranted(ContributionActionInterface::DELETE, $this->status)->thenReturn($isGranted);
+        Phake::when($this->authorizationChecker)->isGranted(ContributionActionInterface::EDIT, $this->status)->thenReturn($isGranted);
         Phake::when($this->groupContext)->hasGroup(Phake::anyParameters())->thenReturn($hasGroup);
         Phake::when($this->translator)->trans(Phake::anyParameters())->thenReturn('trans_display_color');
         Phake::when($this->status)->getDisplayColor()->thenReturn('display_color');
@@ -105,7 +106,7 @@ class StatusTransformerTest extends AbstractBaseTestCase
         $this->assertSame($initialState, $facade->initialState);
         $this->assertSame($isGranted, $facade->allowed);
 
-        if (!$hasGroup) {
+        if (!$hasGroup && $isGranted) {
             $this->assertArrayHasKey('_self_delete', $facade->getLinks());
             $this->assertArrayHasKey('_self_form', $facade->getLinks());
         } else {
@@ -143,7 +144,7 @@ class StatusTransformerTest extends AbstractBaseTestCase
      */
     public function testExceptionTransform()
     {
-        $this->setExpectedException('OpenOrchestra\BaseApi\Exceptions\TransformerParameterTypeException');
+        $this->expectException('OpenOrchestra\BaseApi\Exceptions\TransformerParameterTypeException');
         $this->transformer->transform(Phake::mock('stdClass'));
     }
