--- BackofficeBundle/DependencyInjection/Compiler/AuthorizeStatusChangeCompilerPass.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace OpenOrchestra\BackofficeBundle\DependencyInjection\Compiler;
-
-use OpenOrchestra\BaseBundle\DependencyInjection\Compiler\AbstractTaggedCompiler;
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-
-/**
- * Class AuthorizeStatusChangeCompilerPass
- */
-class AuthorizeStatusChangeCompilerPass extends AbstractTaggedCompiler implements CompilerPassInterface
-{
-    /**
-     * You can modify the container here before it is dumped to PHP code.
-     *
-     * @param ContainerBuilder $container
-     *
-     * @api
-     */
-    public function process(ContainerBuilder $container)
-    {
-        $managerName = 'open_orchestra_backoffice.authorize_status_change_manager';
-        $tagName = 'open_orchestra_backoffice.authorize_status_change.strategy';
-
-        $this->addStrategyToManager($container, $managerName, $tagName);
-    }
-}
