--- BackofficeBundle/Resources/config/authorize_status_change.yml
@@ -1,15 +1,0 @@
-parameters:
-    open_orchestra_backoffice.authorize_status_change_manager.class: OpenOrchestra\BackofficeBundle\StrategyManager\AuthorizeStatusChangeManager
-    open_orchestra_backoffice.authorize_status_change.node.class: OpenOrchestra\Backoffice\AuthorizeStatusChange\Strategies\NodeStrategy
-
-services:
-    open_orchestra_backoffice.authorize_status_change_manager:
-        class: '%open_orchestra_backoffice.authorize_status_change_manager.class%'
-        arguments:
-            - '@security.authorization_checker'
-    open_orchestra_backoffice.authorize_status_change.node:
-        class: '%open_orchestra_backoffice.authorize_status_change.node.class%'
-        arguments:
-            - '@security.authorization_checker'
-        tags:
-            - { name: open_orchestra_backoffice.authorize_status_change.strategy }
