--- WorkflowAdminBundle/Resources/config/authorize_status_change.yml
@@ -1,11 +1,0 @@
-parameters:
-    open_orchestra_workflow_admin.authorize_status_change.workflow_right.class: OpenOrchestra\WorkflowAdminBundle\AuthorizeStatusChange\Strategies\WorkflowRightStrategy
-services:
-    open_orchestra_workflow_admin.authorize_status_change.workflow_right:
-        class: '%open_orchestra_workflow_admin.authorize_status_change.workflow_right.class%'
-        arguments:
-            - '@security.authorization_checker'
-            - '@open_orchestra_model.repository.role'
-            - '@open_orchestra_model.repository.workflow_function'
-        tags:
-            - { name: open_orchestra_backoffice.authorize_status_change.strategy }
